update for 64bit IDE

+++ 3gl/WASListener/Uniface.cpp
@@ -118,9 +118,9 @@ void CUniface::addFileAction(CFileAction const& fileAction)
 
 std::wstring CUniface::s2ws(const std::string& str)
 {
-	const int size_needed = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), NULL, 0);
+	const int size_needed = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), static_cast<int>(str.length()), NULL, 0);
 	std::wstring wstrTo(size_needed, 0);
-	MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), &wstrTo[0], size_needed);
+	MultiByteToWideChar(CP_UTF8, 0, str.c_str(), static_cast<int>(str.length()), &wstrTo[0], size_needed);
 	return wstrTo;
 }
